connected order table with backend

+++ server/common-data-service/src/main/java/com/erp/common/payload/request/DeliveryRequestPayload.java
@@ -5,9 +5,9 @@ import lombok.Data;
 @Data
 public class DeliveryRequestPayload {
     private String code;
-    private Integer value;
+    private long value;
     private String name;
-    private Integer population;
-    private Integer density;
-    private Integer userId;
+    private long population;
+    private long density;
+    private long userId;
 }
